separated logic for displaying the category
